moved files into correct folders position instead of globally available to cause site slow
