+++ frontend_react/src/app/(user)/dashboard/page.tsx
@@ -2,7 +2,7 @@
 
 import Header from "../components/Header";
 import Navbar from "../components/Navbar";
-import { useModules } from "../context/ModulesProvider"; //-- allmodules
+import { useModules } from "../modules/add-modules/context/ModulesProvider"; //-- allmodules
 
 export default function Dashboard() {
   const { modules } = useModules(); //-- allmodules
