+++ frontend_react/src/app/(user)/modules/add-modules/context/ModulesProvider.tsx
@@ -7,7 +7,7 @@ import {
   useState,
   ReactNode,
 } from "react";
-import { FetchAllModules } from "../components/FetchAllModules"
+import { FetchAllModules } from "../../../utils/FetchAllModules"
 
 interface Module {
   moduleCode: string;
