+++ frontend_react/src/app/(user)/utils/CalculateAverageGrade.tsx
