+++ frontend_react/src/app/components/auth/FetchCsrfToken.tsx
